Final commit, everything working
