+++ bode/bode/models/task_relations_actions.py
@@ -1,10 +1,9 @@
-from collections import defaultdict
-
 from sqlalchemy import or_
 
 from bode.app import db
 from bode.models.task_relation import RelationType, TaskRelation
 from bode.resources.task_relations.schemas import DirectedRelationType
+from bode.utils.graph import Graph
 
 
 def get_lhs_related_tasks(task_id, filters=list()):
@@ -55,81 +54,37 @@ def get_relation_types(task_id):
     return {map_to_related_task_schema(relation, task_id) for relation in relations_lhs + relations_rhs}
 
 
-class FindUnion:
-    def __init__(self, uuids):
-        self.parents = dict.fromkeys(uuids, -1)
-        self.ranks = dict.fromkeys(uuids, 0)
-
-    def find(self, x):
-        if self.parents[x] == -1:
-            return x
-
-        self.parents[x] = self.find(self.parents[x])
-        return self.parents[x]
-
-    def union(self, x, y):
-        parents = self.parents
-        ranks = self.ranks
-
-        x_root = self.find(x)
-        y_root = self.find(y)
-
-        if ranks[x_root] > ranks[y_root]:
-            parents[y_root] = x_root
-        elif ranks[x_root] < ranks[y_root]:
-            parents[x_root] = y_root
-        elif x_root != y_root:
-            parents[y_root] = x_root
-            ranks[x_root] += 1
-
-
-class Graph:
-    def __init__(self, V):
-        self.V = V
-        self.graph = defaultdict(list)
-
-    def add_edge(self, u, v):
-        self.graph[u].append(v)
-
-    def compress(self, uuids):
-        self.find_union = FindUnion(uuids)
-        find, union = self.find_union.find, self.find_union.union
-
-        for i in self.graph:
-            for j in self.graph[i]:
-                x = find(i)
-                y = find(j)
-                if x != y:
-                    union(x, y)
-
-        return self.find_union.parents
-
-
 def get_transitive_interchangable_related_tasks(task_id, filters=list()):
     from bode.models.task import Task
 
-    def get_unique_uuids(uuids):
+    def get_distinct_uuids(uuids):
         return list(set(uuids))
+
+    def get_relation_graph(relations):
+        graph = Graph(len(all_interchangable_tasks))
+        uuids = []
+
+        for relation in relations:
+            first_task_id, second_task_id = relation.first_task_id, relation.second_task_id
+            graph.add_edge(str(first_task_id), str(second_task_id))
+            uuids.append(str(first_task_id))
+            uuids.append(str(second_task_id))
+
+        return graph, uuids
+
+    def group_tasks_subsets():
+        task_parent = parents[task_id] if task_id in parents.keys() and parents[task_id] != -1 else task_id
+        tasks_subset = []
+        for key in parents.keys():
+            if task_parent == graph.find_union.find(key):
+                tasks_subset.append(key)
+        return tasks_subset
 
     all_interchangable_tasks = db.session.query(TaskRelation).filter(*filters).all()
 
-    graph = Graph(len(all_interchangable_tasks))
-    uuids = []
-
-    for relation in all_interchangable_tasks:
-        first_task_id, second_task_id = relation.first_task_id, relation.second_task_id
-        graph.add_edge(str(first_task_id), str(second_task_id))
-        uuids.append(str(first_task_id))
-        uuids.append(str(second_task_id))
-
-    parents = graph.compress(get_unique_uuids(uuids))
-
-    interchangable_tasks = []
-    task_parent = parents[task_id] if task_id in parents.keys() and parents[task_id] != -1 else task_id
-
-    for key, val in parents.items():
-        if task_parent == graph.find_union.find(key):
-            interchangable_tasks.append(key)
+    graph, uuids = get_relation_graph(all_interchangable_tasks)
+    parents = graph.union_subsets(get_distinct_uuids(uuids))
+    interchangable_tasks = group_tasks_subsets()
 
     return (
         db.session.query(TaskRelation, Task)
